feat: 오늘의 챌린지 api 추가

+++ challenge/schemas.py
@@ -1,13 +1,12 @@
 from datetime import datetime
-from decimal import Decimal
 from typing import Optional
 
 from ninja import Schema
 
 
 class CoordinateSchema(Schema):
-    latitude: Decimal
-    longitude: Decimal
+    latitude: float
+    longitude: float
 
 
 class ChallengeAcceptSchema(Schema):
